Implement an abstract method for Expenses
