+++ app/src/main/java/android/h/horizon/budget_assistant/expenses_layer/ExpenseActivity.java
@@ -1,0 +1,26 @@
+package android.h.horizon.budget_assistant.expenses_layer;
+
+import android.h.horizon.budget_assistant.R;
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+
+public abstract class ExpenseActivity extends AppCompatActivity {
+    protected abstract void setActivityTitle();
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_expense);
+        setTitle("Food");
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        Fragment fragment = fragmentManager.findFragmentById(R.id.expense_fragment_container);
+        if (fragment == null) {
+            fragment = new FoodFragment();
+            fragmentManager.beginTransaction()
+                    .add(R.id.expense_fragment_container, fragment)
+                    .commit();
+        }
+    }
+}
