+++ app/src/main/java/android/h/horizon/budget_assistant/expenses_layer/FoodActivity.java
@@ -7,19 +7,23 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 
-public class FoodActivity extends AppCompatActivity {
+public class FoodActivity extends ExpenseActivity {
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_expense);
+    protected void setActivityTitle() {
         setTitle("Food");
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        Fragment fragment = fragmentManager.findFragmentById(R.id.expense_fragment_container);
-        if (fragment == null) {
-            fragment = new FoodFragment();
-            fragmentManager.beginTransaction()
-                    .add(R.id.expense_fragment_container, fragment)
-                    .commit();
-        }
     }
+//    @Override
+//    protected void onCreate(Bundle savedInstanceState) {
+//        super.onCreate(savedInstanceState);
+//        setContentView(R.layout.activity_expense);
+//        setTitle("Food");
+//        FragmentManager fragmentManager = getSupportFragmentManager();
+//        Fragment fragment = fragmentManager.findFragmentById(R.id.expense_fragment_container);
+//        if (fragment == null) {
+//            fragment = new FoodFragment();
+//            fragmentManager.beginTransaction()
+//                    .add(R.id.expense_fragment_container, fragment)
+//                    .commit();
+//        }
+//    }
 }
